refact: moved types in another mod
