+++ src/ast/cdl/mod.rs
@@ -1,4 +1,5 @@
 mod actions_and_functions;
+mod types;
 mod entity_term;
 mod field_term;
 mod from_term;
@@ -9,9 +10,9 @@ mod param_term;
 mod path_term;
 mod returns_term;
 mod service_term;
-mod type_term;
 
 pub use actions_and_functions::*;
+pub use types::*;
 pub use entity_term::*;
 pub use field_term::*;
 pub use from_term::*;
@@ -22,4 +23,3 @@ pub use param_term::*;
 pub use path_term::*;
 pub use returns_term::*;
 pub use service_term::*;
-pub use type_term::*;
