+++ src/ast/cdl/module_term.rs
@@ -3,7 +3,7 @@ use crate::ast::*;
 use super::super::import::ImportTerm;
 use super::entity_term::EntityTerm;
 use super::service_term::ServiceTerm;
-use super::type_term::TypeTerm;
+use super::TypeTerm;
 use ast_term_derive::ASTTerm;
 
 #[derive(ASTTerm, PartialEq, Eq, Debug, Clone)]
