+++ src/ast/cdl/types/type_term.rs
@@ -1,4 +1,4 @@
-use super::name_term::NameTerm;
+use super::super::NameTerm;
 use ast_term_derive::ASTTerm;
 
 #[derive(ASTTerm, PartialEq, Eq, Debug, Clone)]
