+++ src/parser/single_module/grammar/cdl/types/grammar.lalrpop
@@ -1,0 +1,8 @@
+Type: TypeTerm = {
+  "define" "type" <name:Name> ":" <resolved_type_name:Name> ";" => {
+    TypeTerm::new(Box::new(name), Box::new(resolved_type_name))
+  },
+  "type" <name:Name> ":" <resolved_type_name:Name> ";" => {
+    TypeTerm::new(Box::new(name), Box::new(resolved_type_name))
+  }
+}
